requirements.txt
###### Requirements ######`
numpy
scipy
matplotlib
termcolor
argparse
munkres >= 1.0.12
pickle5
nuitka
PyYAML
natsort
azure-batch
azure-mgmt-batch
azure-identity
